# App-Monopatin-Electrico/microEstacion/src/main/resources/swagger.yaml
openapi: 3.0.0
info:
  title: API de Gestión de Estaciones
  version: 1.0.0
  description: |
    API para administrar estaciones, incluyendo creación, actualización, eliminación y consulta de estaciones.

servers:
  - url: http://localhost:8050/
    description: Servidor local de desarrollo

paths:
  /estacion:
    get:
      summary: Obtener todas las estaciones
      responses:
        '200':
          description: Lista de estaciones obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EstacionResponseDTO'
    post:
      summary: Crear una nueva estación
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EstacionRequestDTO'
      responses:
        '200':
          description: Estación creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstacionResponseDTO'

  /estacion/{id}:
    get:
      summary: Obtener una estación por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string  # Cambio de integer a string
      responses:
        '200':
          description: Estación encontrada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstacionResponseDTO'
        '404':
          description: Estación no encontrada
    put:
      summary: Actualizar una estación
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string  # Cambio de integer a string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EstacionRequestDTO'
      responses:
        '200':
          description: Estación actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstacionResponseDTO'
        '404':
          description: Estación no encontrada
    delete:
      summary: Eliminar una estación
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string  # Cambio de integer a string
      responses:
        '204':
          description: Estación eliminada exitosamente
        '404':
          description: Estación no encontrada

components:
  schemas:
    EstacionRequestDTO:
      type: object
      properties:
        id:
          type: string  # Cambio de integer a string
          nullable: true
        latitud:
          type: number
          format: double  # Cambio a tipo double
          nullable: false
        longitud:
          type: number
          format: double  # Cambio a tipo double
          nullable: false
      required:
        - latitud
        - longitud

    EstacionResponseDTO:
      type: object
      properties:
        id:
          type: string  # Cambio de integer a string
        latitud:
          type: number
          format: double  # Cambio a tipo double
        longitud:
          type: number
          format: double  # Cambio a tipo double